qdoc: Reintroduce Q_OS_* defines into the global documentation config
This was dropped by commit 740ff1b8 in preparation for clang-based
C++ parser in qdoc. However, the old (current) parser still needs it,
as some of the functions to document are wrapped with those macros.

It does not interfere with the clang parser anyway; unlike the old
parser, clang will not resolve the wildcards in macros.

Change-Id: I3135a263e8eb200a11b1e918c776e3a2c3ec5fb6
Reviewed-by: Martin Smith <martin.smith@qt.io>