Fix MAYA-72256 Torn-off menu on small screen can freeze Maya
From 89799748f28bb9718d8b5da68c00d495a18c4a83 Mon Sep 17 00:00:00 2001
From: Dongmei Wang <dongmei.wang@qt.io>
Date: Wed, 1 Feb 2017 20:43:00 -0800
Subject: [PATCH 1/2] QMenu: Fix margins related display issues

Currently the contents margins and the menu paddings are not considered
for calculating the menu size, the positions and the size of tear-off
bar, scrollers and the positions of the menu items when scrolling the
menu, which results in the following problems when valid contents
margins and/or menu paddings are set:

- The tear off area is displayed in a wrong position. The mouse events
are not handled correctly in the tear off area. For example, when you
click in the tear off area, the menu should be torn off but nothing
happens
- For a multi-column menu, the menu width is not calculated correctly
- For a scrollable menu,
  - the menu width is not calculated correctly
  - the menu items are not displayed in correct positions
  - the scrollers are not displayed in correct positions
  - menu items are displayed on the area of borders and margins when
    scrolling the menu
  - the last menu item is not displayed above the bottom of the content
    area when scrolling the menu to the end.

The changes are to fix the problems above.

Change-Id: I7931e1088dff0029f2d4825e2aa34b4e32fdccd9

From 58e6f661c01258a5451d587c1e8530c586c97aa5 Mon Sep 17 00:00:00 2001
From: Dongmei Wang <dongmei.wang@qt.io>
Date: Thu, 17 Nov 2016 14:00:30 -0800
Subject: [PATCH 2/2] QMenu: Fix torn-off menu display crash issue

When tearing off either a non-scrollable multi-colume menu
or a scrollable menu, displaying the torn-off menu crashes.
The root cause is when the torn-off menu is created, the
tear-off menu's style, margins and other attributes are not
set to it. The patch is to ensure the torn-off menu has
the same attributes as the tear-off menu does and set the
torn-off menu with a correct menu size.

Change-Id: Icea45f149ea8792671af4a62e62cad6ee01a1f95