Android: Allow deploying Qt apps as system apps
The main app shared library and shared dependencies are not automatically
deployed when the apk is placed in /system/app or /system/priv-app. In
such cases, we need to place these libraries in /system/lib and tell
QtLoader to load them from here. It is possible to specify a custom
library path in AndroidManifest.xml.

[ChangeLog][Android][QtLoader] Enabled loading shared libraries from
/system/lib or a custom path specified with the
android.app.system_libs_prefix metadata variable in AndroidManifest.xml.
This allows deploying Qt apps as Android system apps.

Change-Id: I8388e91a53475b06a027467face45c08f096fbf8
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>
Reviewed-by: Anton Kudryavtsev <antkudr@mail.ru>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>